+++ packages/backend/tests/program/sortPrograms.test.ts
@@ -1,8 +1,8 @@
 import { sortPrograms } from '@tee/backend/src/program/domain/sortPrograms'
-import { ProgramAidType } from '@tee/web/src/types'
 import { QuestionnaireRoute } from '@tee/common/src/questionnaire/types'
 import { makeProgramHelper } from './testing'
 import { Program } from '../../src/program/domain/types'
+import { ProgramAidType } from '@tee/common/src/program/types'
 
 const makeProgram = (id: string, nature: ProgramAidType, cost: string = '') =>
   makeProgramHelper({
