+++ packages/common/src/program/types.ts
@@ -1,0 +1,7 @@
+export enum ProgramAidType {
+  acc = 'accompagnement',
+  train = 'formation',
+  fund = 'financement',
+  loan = 'prêt',
+  tax = 'avantage fiscal'
+}
