+++ packages/web/src/types/programTypes.ts
@@ -1,12 +1,6 @@
 // FOR AID PROGRAMS
 
-export enum ProgramAidType {
-  acc = 'accompagnement',
-  train = 'formation',
-  fund = 'financement',
-  loan = 'prêt',
-  tax = 'avantage fiscal'
-}
+import type { ProgramAidType } from '@tee/common/src/program/types'
 
 export interface ProgramData {
   id: string
@@ -21,7 +15,6 @@ export interface ProgramData {
   'opérateur de contact': string
   'autres opérateurs'?: string[]
   "nature de l'aide": ProgramAidType
-
   "coût de l'accompagnement"?: string
   "durée de l'accompagnement"?: string
   'montant du financement'?: string
